+++ CameraControl.Devices/ICameraDevice.cs
@@ -59,7 +59,7 @@ namespace CameraControl.Devices
         bool HostMode { get; set; }
 
         PropertyValue<int> FNumber { get; set; }
-        PropertyValue<int> IsoNumber { get; set; }
+        PropertyValue<long> IsoNumber { get; set; }
         PropertyValue<long> ShutterSpeed { get; set; }
         PropertyValue<long> WhiteBalance { get; set; }
         PropertyValue<uint> Mode { get; set; }
